feat(conductor)!: implement chain ID checks (#1482)
## Summary
Implemented chain ID validation for both the sequencer and Celestia
readers in Conductor.

## Background
This is to safeguard against connecting to an incorrect network.

This change was previously blocking on a switch to solely using gRPC for
communication with sequencer, but after offline discussion it was
decided to implement the chain ID check anyways using the HTTP client.

## Changes
- Added sequencer and Celestia chain ID environment variables.
- Added chain ID checks to the initialization of both the sequencer and
Celestia readers.
- Changed conductor `run_until_stopped()` to return with an error so
that it will not fail silently.

## Testing
Added 2 blackbox tests to ensure the conductor exits with the proper
error in the event of both a Celestia and sequencer chain ID mismatch.

## Breaking Changelist
- Added `ASTRIA_CONDUCTOR_SEQUENCER_CHAIN_ID` and
`ASTRIA_CONDUCTOR_CELESTIA_CHAIN_ID` environment variables, without
which conductor will not function.

## Related Issues
closes #986

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>