chore(sequencer): simplify boolean expressions in `transaction container` (#1595)
## Summary
Simplified logical statements in
`transaction_priority_comparisons_should_be_consistent_nonce_diff()`.

## Background
Previously there was an allow for `clippy::nonminimal_bool`. The
reasoning behind it was to match documented behavior. This change is
meant to explicitly state the expected behavior while still simplifying
the boolean expressions. This is in response to this comment:
https://github.com/astriaorg/astria/pull/1561#discussion_r1778613519

## Changes
- Simplified boolean expressions and moved the non-simplified versions
to the comments where applicable to provide context on the documented
behavior.

## Testing
Passing all tests

## Related Issues
closes #1583