feat(proto): add bundle and optimistic block apis (#1519)
## Summary
This adds the protobuf definitions for optimistic block stream and the
bundle stream.

## Background
The Auctioneer will auction off a bundle slot which will be placed at
the top of the rollup block deterministically during execution. The APIs
in this PR are used to drive the bundle auction (detailed in [this
document](https://www.notion.so/astria-org/20240919-Status-Update-1066bd31a90c80b68bb2ea9b898645d0)).
At a high level:

The Auctioneer receives blocks optimistically from the sequencer in
order to maximize the auction duration.

While the auction is running, the Auctioneer receives orders from the
rollup node that have been validated against the optimistic block.

After a block is committed, the Auctioneer will submit the highest
paying bundle for inclusion by the sequencer.

### Optimistic Block Stream
1. The Sequencer will stream blocks to the Auctioneer optimistically,
i.e. before they are finalized (ProcessProposal).
2. The Sequencer will also stream block commitments to the Auctioneer,
which mark blocks as final thanks to CometBFT's single slot finality.

### Bundle Stream
1. The Auctioneer will stream optimistic blocks to the rollup node for
execution. The rollup will execute the block and return the resulting
header info.
2. The rollup node will stream bundles (indexed by parent block hash) to
the Auctioneer.

## Changes
- Add the bundle service to the composer apis
- Add the optimistic block service to the sequencer block apis

## Breaking Changelist
- Bulleted list of breaking changes, any notes on migration. Delete
section if none.

## Related Issues
closes https://github.com/astriaorg/astria/issues/1553