refactor(sequencer): remove unused asset storage variant (#1704)
## Summary
Removes an unused variant of the internal storage implementation of
sequencer asset component.t

## Background
As part of https://github.com/astriaorg/astria/pull/1647 fees were moved
from the assets to the fees component. The present variant was
overlooked.

## Changes
- Remove a storage variant from the assets component.

## Testing
Not applicable. Dead code.

## Breaking Changelist
This is not a breaking change because a) the variant was never used
(there was not even a theoretic way to write it after
https://github.com/astriaorg/astria/pull/1647) and b) no variants
existed after the variant. If they had, this change would lead to a
shift in the variants, causing breakage.

---------

Co-authored-by: Fraser Hutchison <fraser@astria.org>