feat(core, sequencer)!: make native asset optional (#1703)
## Summary
Makes setting a native asset optional.

## Background
Native assets are a remnant of sequencer that is now deeply entrenched
in how it runs (whitebox) tests. While we don't want to set a native
assets, in mainnet, it is too big a lift to remove them at this point.

## Changes
- Change genesis validation to permit empty native assets.
- Update account chain init to only require native assets if accounts
are set.
- Remove native asset lookup from querying assets (this can be done via
looking up all ibc assets.

## Testing
The absence of native assets is not tested as of this patch. All tests
still function with native assets set.

A proper test would follow these steps:
1. Spin up a test net without native assets, accounts, or permitted fee
assets set
2. Bridge in an IBC asset via an ics20 transfer
3. Set a permitted fee asset to the bridged asset
4. Execute transactions.


## Breaking Changelist
This is marked as breaking since a node with before this change would
reject an empty native asset in its genesis while a node after this
patch would initialize successfully.

However, this is mild breakage because both nodes would not form a
network in the first place.

---------

Co-authored-by: quasystaty <ido@astria.org>
Co-authored-by: Jordan Oroshiba <jordan@astria.org>