chore(sequencer): init allowed assets in fees component (#1730)
## Summary
Moves the initialization of allowed assets from the overall app to the
fees component.

## Background
https://github.com/astriaorg/astria/pull/1647 moved fee asset state
handling to the new `fees` component. It was overlooked that writing
allowed fee assets should also happen in its init chain logic.

## Changes
- Move initialization of app init state to the fees component.

## Testing
Nothing has changed in terms of how the state is represented. Tests that
rely on init chain still work.

## Breaking Changelist
This is not a breaking change because the state after this change looks
the same.