chore(proto): reserved only for what would be breaking change (#1684)
## Summary
We have been using reserved to highlight gaps, but reserved should
ideally only be utilized for reserving previously used identifiers.

## Changes
- removes reserved numbers that using were not actual breaking change
- updates workflow to allow override and to always run all breaking
checks even if previous fails

## Breaking Changes
This has no real breaking changes, hence allowing override. The removal
of the reserved makes it so that we don't have to do this in the future
when adding new fields.