fix(charts): genesis template (#1737)
## Summary
changes genesis to allow configuring each fee component, also make
`nativeAssetBaseDenomination ` an optional field
## Background
Fees can now be added after genesis, meaning we no longer have to define
them all in it.
Also in case `nativeAssetBaseDenomination` is unset, it will not be used
when initiating the state and thus should not appear in `genesis.json.
## Changes
- added an if statement to remove the field in case it is nil
- can now set / unset all fees except `feeChange` in genesis

## Testing
locally with helm template command