fix(bridge-withdrawer, cli)!: use decimals from erc20 contract (#1762)
## Summary
Updates CLI and bridge withdrawer to read the decimals off the contract
when it is an erc20, to get the accurate decimals to utilize during
withdrawal.

## Background
Previously it was assumed that all erc20 contracts would have 18
decimals this is not guaranteed. #1760 is a similar fix but relies on
configuring the withdrawer correctly. This version automatically reads
off of the contract instead if it is an erc20.

## Changes
- Read decimals off contract if  or fallback to native prevision.

## Testing
Ran against Flame to ensure it gets decimals correct on 6 decimal
contract + CI tests.

## Changelogs
Changelogs updated.

---------

Co-authored-by: Richard Janis Goldschmidt <github@aberrat.io>