chore(spec): update bridge spec (#1705)
## Summary
Updated the bridge spec to add more details on some of the data
structures and link to the up-to-date actions.

## Background
With the preparation for mainnet, some of the action names and locations
have changed. The spec needed updating to reflect that.

## Related Issues
Link any issues that are related, prefer full github links.

closes <!-- list any issues closed here -->