chore(sequencer): consolidate all action handling to one module (#1759)
## Summary
Moved all action and transaction handling to single module.

## Background
Previously, all implementations of `ActionHandler` were done separately
in components which corresponded to their given actions (or groups of
actions, such as in `bridge`). While this may make sense from an
action-centered point of view, it is confusing to an outsider to have to
look many different places to find where stateless/stateful checks and
execution occur. This PR consolidates all of the implementations as well
as their testing.

## Changes
- Created new `action_handler` module and moved the corresponding
`ActionHandler` trait as well as all of its `impl`s into this module.
- Renamed some moved tests for clarity.
- Categorized tests by action in submodule `tests`.
- Categorized impls by action in submodule `impls`.

## Testing
Passing all tests.

## Changelogs
Changelogs updated

## Related Issues
closes #1657