chore(core, sequencer): add names to all snapshot tests (#1690)
## Summary
Added names to all snapshot tests which did not have them.

## Background
If a certain test had multiple snapshot assertions, they would be named
based on their ordering. Hence any reordering, removal, or addition of
the snapshots would break the test, which should not be the case. This
change aims to fix this problem as well as establish a best practice of
naming snapshots moving forward.

## Changes
- Added names to all snapshot tests and regenerated all snapshots not
within Conductor (due to code freeze)

## Testing
Passing all tests.

## Related Issues
closes #1656