chore(bridge-contracts): add snapshot test for memo (#1753)
## Summary
This adds a snapshot test for the memo field being populated from the
withdrawal info.

## Background
The new test should serve to emphasize that the schema of the memo field
is part of the API and changes to this could well be breaking.

## Changes
- Moved some code unchanged inside the existing `memo_to_json` and
removed the generic parameter.
- Added snapshot test.

## Testing
This _is_ a test.

## Changelogs
No updates required - this is only adding a test.

## Related Issues
Closes #1752.

---------

Co-authored-by: noot <36753753+noot@users.noreply.github.com>