chore: add tests for ABCI error codes (#1754)
## Summary
Added tests for the ABCI error codes.

## Background
The new snapshot test should serve to emphasize that the error codes are
part of the API and changes here would be breaking.

The other test should help avoid accidental duplication of a value.

## Changes
- Added a snapshot test.
- Added a test for unique values of error codes.

## Testing
These _are_ tests.

## Changelogs
No updates required - this is only adding unit tests.

## Related Issues
Closes #1634.

Co-authored-by: noot <36753753+noot@users.noreply.github.com>