fix(sequencer): remove unwrap from app utilized mempool logic  (#1772)
## Summary
This PR changes the mempool's `builder_queue()` to be infallible. The
code previously would return an error if the nonce fetch from the
database failed. Now it handles the error case by using the pending's
lowest nonce for the account as an educated guess of what nonce to use
for transaction priority construction. This is okay since
`prepare_proposal()`'s logic will just reject invalid nonces if the
mempool's state is out of date due to a bug.

## Background
We shouldn't have non-existential issues in the mempool cause failures
in the sequencer.

## Changes
- Changed mempool `builder_queue()` to be infallible with reasonable
fallbacks.

## Testing
Manually changed the code path to only use the new logic and watched all
tests pass except for those explicitly testing the jitter between the
mempool and database (which shouldn't happen if the rest of the code
works). The tests in those scenarios only fail due to unmet assertions
and not because of panics.

## Changelogs
No updates required.


## Related Issues
closes #1769

---------

Co-authored-by: Fraser Hutchison <fraser@astria.org>